+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Changed
 - New colour theme for `?greyline`.
-
+### Fixed
+- The result of `?cond` was cached by discord and would get out of date.
 
 ## [2.5.1] - 2020-12-10
 ### Fixed